data_build/requirements.txt
SQLAlchemy==1.3.3
sqlalchemy-fdw==0.3.0
marshmallow-sqlalchemy==0.16.2
zstandard==0.11.0
pyhumps==1.2.2